Fix: menu mobile refactor
